Refactored rotes, requests  now we don't need to pass userId anymore

+++ src/main/kotlin/routes/CommentRoute.kt
@@ -7,7 +7,6 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import pw.coding.data.requests.CreateCommentRequest
 import pw.coding.data.requests.DeleteCommentRequest
-import pw.coding.data.requests.DeletePostRequest
 import pw.coding.data.responses.BasicApiResponse
 import pw.coding.service.CommentService
 import pw.coding.service.LikeService
@@ -16,46 +15,40 @@ import pw.coding.util.ApiResponseMessages
 import pw.coding.util.QueryParams
 
 fun Route.createComment(
-    commentService: CommentService,
-    userService: UserService
-){
+    commentService: CommentService
+) {
     authenticate {
         post("/api/comment/create") {
             val request = call.receiveNullable<CreateCommentRequest>() ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
+            when (commentService.createComment(request, call.userID)) {
+                CommentService.ValidationEvent.ErrorFieldEmpty ->
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.FIELDS_BLANK
+                        )
+                    )
 
-            ifEmailBelongsToUSer(
-                call= call,
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ){
-                when(commentService.createComment(request)){
-                    CommentService.ValidationEvent.ErrorFieldEmpty ->
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.FIELDS_BLANK
-                            )
+                CommentService.ValidationEvent.ErrorCommentTooLong ->
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.COMMENT_TOO_LONG
                         )
-                    CommentService.ValidationEvent.ErrorCommentTooLong ->
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = ApiResponseMessages.COMMENT_TOO_LONG
-                            )
+                    )
+
+                CommentService.ValidationEvent.Success ->
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true
                         )
-                    CommentService.ValidationEvent.Success ->
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true
-                            )
-                        )
-                }
+                    )
             }
         }
     }
@@ -64,9 +57,9 @@ fun Route.createComment(
 
 fun Route.getCommentsForPost(
     commentService: CommentService
-){
+) {
     authenticate {
-        get("/api/comment/get"){
+        get("/api/comment/get") {
             val postId = call.pathParameters[QueryParams.PARAM_POST_ID] ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@get
@@ -80,29 +73,28 @@ fun Route.getCommentsForPost(
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
-){
-   authenticate {
-       delete("/api/comment/delete"){
-           val request = call.receiveNullable<DeleteCommentRequest>() ?: kotlin.run {
-               call.respond(HttpStatusCode.BadRequest)
-               return@delete
-           }
+) {
+    authenticate {
+        delete("/api/comment/delete") {
+            val request = call.receiveNullable<DeleteCommentRequest>() ?: kotlin.run {
+                call.respond(HttpStatusCode.BadRequest)
+                return@delete
+            }
 
-           ifEmailBelongsToUSer(
-               call=call,
-               userId = request.userId,
-               validateEmail = userService::doesEmailBelongsToUserId
-           ){
-               val deleteComment = commentService.deleteComment(request.commentId)
-               if(deleteComment){
-                   likeService.deleteLikesForParent(request.commentId)
-                   call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
-               }else{
-                   call.respond(HttpStatusCode.OK, BasicApiResponse(successful = false))
-               }
-           }
-       }
-   }
+            val comment = commentService.getCommentById(request.commentId)
+            if(comment?.userId != call.userID){
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+
+            val commentDeleted = commentService.deleteComment(request.commentId)
+            if (commentDeleted) {
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
+            } else {
+                call.respond(HttpStatusCode.OK, BasicApiResponse(successful = false))
+            }
+        }
+    }
 }
